fix(intellij-plugin): fix code review issues
